Minor updates and restructuring the distribution. See NEWS and ChangeLog for details.
2001/06/01 02:07:23-00:00 !antona
It has been a long time since last commit. At moment have done a lot of work
on mkntfs but also at the moment ntfsfix and ntfsdump_logfile and libntfs are
broken. Basically only mkntfs works and that is not complete either.

2001/02/02 00:30:25-00:00 !antona
Had to change symbolic links to files to get cvs to accept them.

(Logical change 1.5)