whitespace and include guards
2002/07/03 21:56:01-00:00 !antona
Updates

2002/07/03 14:38:40-00:00 !antona
Cleanup the build process a bit.

2002/07/02 23:47:11-00:00 !antona
Global replacement of __[su]{8,16,32,64} with [su]{8,16,32,64} and layout.h define it.

2002/04/21 01:26:39-00:00 !antona
Cleanup/streamline include file dependencies.

2002/04/18 18:15:46-00:00 !antona
Define API for bootsect.[ch]: is_boot_sector_ntfs().

2002/04/16 12:13:53-00:00 !antona
New API function mft.[ch]::read_file_record(). Also some cleanups.

2001/06/16 19:59:51-00:00 !antona
Update ntfsdump_logfile and dumplog and the relevant textfiles.

2001/06/16 00:06:17-00:00 !antona
dumplog now decodes the first 0x5000 bytes of the logfile given on the command
line as fully as possible with the current knowledge of the logfile structures.

2001/06/15 16:45:12-00:00 !antona
New log dumper to operate on a file rather than a partition and it is more
advanced and corrected compared to ntfsdump_logfile.
Also add a file that will do the creation of the logfile structures for mkntfs.

(Logical change 1.5)