Forgot that one.
2001/06/10 18:30:16-00:00 !antona
Preparations for the 0.1.0 release. Updating documentation and information.
Updating rpm generation and added mkntfs man page which currently is not
accurate as it shows all the mke2fs options in it while mkntfs doesn't actually
accept any options except for device at all but it is a good starting point.

2001/02/02 01:15:17-00:00 !antona
Updated documentation and other text files. Preparing for initial release.

2001/02/02 00:30:25-00:00 !antona
Had to change symbolic links to files to get cvs to accept them.

(Logical change 1.5)