configure.ac: Respect 'with_fuse' user setting on non-Linux/Solaris.
Previously the configure script simply disregarded the value of the
'with_fuse' flag on non-Linux/Solaris systems. This is not unreasonable
since the 'internal' option doesn't work on those systems, however in
some situations we want to specify the libfuse CFLAGS/LDFLAGS manually
and bypass the pkg-config check with '--with-fuse=none' and this was not
possible with this logic. So add a special test to bypass the automatic
determination when the user has specified a 'with_fuse' value.