Clarify license of boot.c to be public domain and not GPL-3
The licensing of boot.c currently makes mkntfs special as it's the only
tool which is subject to the GPL-3 terms.

However, the only GPL-3 bits originally come from a section of the
dosfstools mkfs.fat.c that has been later clarified to be under
public domain:

https://github.com/dosfstools/dosfstools/commit/3ce32fa90d1d97aa7070dad3e98db4a8b8b07fc3

Reflect this in the ntfs-3g version as well, to let people who have
issues (for whatever reason) with the GPL-3 to use ntfs-3g without
any concern.

Signed-off-by: Emanuele Aina <emanuele.aina@collabora.com>